Library

LaTeXFigures.Position — Type

Position(n)

Represent the position of the figure, available values are here, top, bottom, page, override, Here.

LaTeXFigures.Figure — Type

Figure(path, caption, label, position, centering, options)

Represent a single figure.

LaTeXFigures.Subfigure — Type

Subfigure(path, w, h, caption, label, position, centering, options)

Represent a subfigure within a figure.

LaTeXFigures.TwoSubfigures — Type

TwoSubfigures(a, b, caption, label, position, centering, hfill)

Represent a figure with two subfigures a and b.

LaTeXFigures.latexformat — Function

latexformat(figure::Figure; indent=' '^4, newline='\n')
latexformat(figure::Subfigure; indent=' '^4, newline='\n')
latexformat(figure::TwoSubfigures; indent=' '^4, newline='\n')

Return a formatted LaTeX string representing the figure.